comentario

+++ src/Interfaces/IEvaluarMuestra.java
@@ -6,5 +6,5 @@ package Interfaces;
  * @author erley
  */
 public interface IEvaluarMuestra {
-    int[][] EvlauarMuestra(int [] muestra);
+    double [][] EvlauarMuestra(int [] muestra);
 }
